Created LightCurve base class and separate StructureFunctionLightCurve child class. Updated references and tests.
