+++ src/lsstseries/analysis/__init__.py
@@ -1,2 +1,3 @@
+from .light_curve import LightCurve  # noqa
 from .stetsonj import *  # noqa
 from .structurefunction2 import *  # noqa
