+++ src/lsstseries/analysis/structure_function/base_calculator.py
@@ -5,7 +5,7 @@ import numpy as np
 from scipy.stats import binned_statistic
 
 from lsstseries.analysis.structure_function.base_argument_container import StructureFunctionArgumentContainer
-from lsstseries.analysis.structure_function.light_curve import StructureFunctionLightCurve
+from lsstseries.analysis.structure_function.sf_light_curve import StructureFunctionLightCurve
 
 
 class StructureFunctionCalculator(ABC):
