+++ src/lsstseries/analysis/structure_function/macleod_2012/calculator.py
@@ -4,7 +4,6 @@ import numpy as np
 
 from lsstseries.analysis.structure_function.base_argument_container import StructureFunctionArgumentContainer
 from lsstseries.analysis.structure_function.base_calculator import StructureFunctionCalculator
-from lsstseries.analysis.structure_function.light_curve import StructureFunctionLightCurve
 
 # MacLeod et al. 2012, Erratum 2014ApJ...782..119M
 CONVERSION_TO_SIGMA = 0.74
