+++ tests/lsstseries_tests/structure_function_calculators/test_base_calculator.py
@@ -3,7 +3,7 @@ import pytest
 
 from lsstseries.analysis.structure_function.base_argument_container import StructureFunctionArgumentContainer
 from lsstseries.analysis.structure_function.basic.calculator import BasicStructureFunctionCalculator
-from lsstseries.analysis.structure_function.light_curve import StructureFunctionLightCurve
+from lsstseries.analysis.structure_function.sf_light_curve import StructureFunctionLightCurve
 
 
 def test_dt_bins():
