+++ tests/lsstseries_tests/structure_function_calculators/test_macleod_2012_calculator.py
@@ -4,7 +4,7 @@ from lsstseries.analysis.structure_function.macleod_2012.calculator import (
     Macleod2012StructureFunctionCalculator,
 )
 from lsstseries.analysis.structure_function.base_argument_container import StructureFunctionArgumentContainer
-from lsstseries.analysis.structure_function.light_curve import StructureFunctionLightCurve
+from lsstseries.analysis.structure_function.sf_light_curve import StructureFunctionLightCurve
 
 
 def test_basic_calculation():
